+++ components/auth/signup/SignupForm.style.ts
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const SignUpFormContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+
+  margin-top: 16px;
+  height: 100%;
+`;
+
+export const StepContainer = styled.div`
+  display: flex;
+  gap: 4px;
+`;
